+++ tests/gui/test_03_data_overview.py
@@ -1,7 +1,7 @@
 from streamlit.testing.v1 import AppTest
 from pathlib import Path
 from unittest.mock import MagicMock, patch
-from .test_helper import create_dataset_alphapept
+from .conftest import create_dataset_alphapept
 import pandas as pd
 from io import BytesIO
